refactor(carTest): 테스트 코드 위치 변경 및 CarGroupsTest 클래스 추가
